requirements.txt
flask
flask-cors
pillow
numpy
opencv-python
easyocr
ultralytics
pandas
openpyxl